+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/PostgresLedgerExtenderService.cs
@@ -1136,10 +1136,9 @@ UPDATE pending_transactions
             var mostRecentEntityResourceAggregateHistory = await readHelper.MostRecentEntityResourceAggregateHistoryFor(vaultSnapshots, token);
             var mostRecentEntityResourceAggregatedVaultsHistory = await readHelper.MostRecentEntityResourceAggregatedVaultsHistoryFor(vaultChanges, token);
             var mostRecentEntityResourceVaultAggregateHistory = await readHelper.MostRecentEntityResourceVaultAggregateHistoryFor(vaultSnapshots, token);
-            var mostRecentNonFungibleIdStoreHistory = await readHelper.MostRecentNonFungibleIdStoreHistoryFor(nonFungibleIdChanges, token);
             var mostRecentResourceEntitySupplyHistory = await readHelper.MostRecentResourceEntitySupplyHistoryFor(resourceSupplyChanges, token);
             var mostRecentEntityNonFungibleVaultHistory = await readHelper.MostRecentEntityNonFungibleVaultHistory(vaultSnapshots.OfType<NonFungibleVaultSnapshot>().ToList(), token);
-            var existingNonFungibleIdData = await readHelper.ExistingNonFungibleIdDataFor(nonFungibleIdChanges, vaultSnapshots.OfType<NonFungibleVaultSnapshot>().ToList(), token);
+            var existingNonFungibleIdData = await readHelper.ExistingNonFungibleIdDefinitionFor(nonFungibleIdChanges, vaultSnapshots.OfType<NonFungibleVaultSnapshot>().ToList(), token);
 
             await entityMetadataProcessor.LoadDependencies();
             await entitySchemaProcessor.LoadDependencies();
@@ -1159,8 +1158,7 @@ UPDATE pending_transactions
             var entityResourceAggregateHistoryCandidates = new List<EntityResourceAggregateHistory>();
             var entityResourceAggregatedVaultsHistoryToAdd = new List<EntityResourceAggregatedVaultsHistory>();
             var entityResourceVaultAggregateHistoryCandidates = new List<EntityResourceVaultAggregateHistory>();
-            var nonFungibleIdStoreHistoryToAdd = new Dictionary<NonFungibleStoreLookup, NonFungibleIdStoreHistory>();
-            var nonFungibleIdDataToAdd = new List<NonFungibleIdData>();
+            var nonFungibleIdDataToAdd = new List<NonFungibleIdDefinition>();
             var nonFungibleIdLocationHistoryToAdd = new List<NonFungibleIdLocationHistory>();
             var nonFungibleIdsMutableDataHistoryToAdd = new List<NonFungibleIdDataHistory>();
 
@@ -1186,9 +1184,9 @@ UPDATE pending_transactions
                     new NonFungibleIdLookup(e.ReferencedResource.DatabaseId, e.NonFungibleId),
                     _ =>
                 {
-                    var ret = new NonFungibleIdData
+                    var ret = new NonFungibleIdDefinition
                     {
-                        Id = sequences.NonFungibleIdDataSequence++,
+                        Id = sequences.NonFungibleIdDefinitionSequence++,
                         FromStateVersion = e.StateVersion,
                         NonFungibleResourceEntityId = e.ReferencedResource.DatabaseId,
                         NonFungibleId = e.NonFungibleId,
@@ -1199,42 +1197,16 @@ UPDATE pending_transactions
                     return ret;
                 });
 
-                var nonFungibleIdStore = nonFungibleIdStoreHistoryToAdd.GetOrAdd(
-                    new NonFungibleStoreLookup(e.ReferencedResource.DatabaseId, e.StateVersion),
-                    _ =>
-                {
-                    IEnumerable<long> previousNonFungibleIdDataIds = mostRecentNonFungibleIdStoreHistory.TryGetValue(e.ReferencedResource.DatabaseId, out var value)
-                        ? value.NonFungibleIdDataIds
-                        : Array.Empty<long>();
-
-                    var ret = new NonFungibleIdStoreHistory
-                    {
-                        Id = sequences.NonFungibleIdStoreHistorySequence++,
-                        FromStateVersion = e.StateVersion,
-                        NonFungibleResourceEntityId = e.ReferencedResource.DatabaseId,
-                        NonFungibleIdDataIds = new List<long>(previousNonFungibleIdDataIds),
-                    };
-
-                    mostRecentNonFungibleIdStoreHistory[e.ReferencedResource.DatabaseId] = ret;
-
-                    return ret;
-                });
-
                 nonFungibleIdsMutableDataHistoryToAdd.Add(
                     new NonFungibleIdDataHistory
                 {
                     Id = sequences.NonFungibleIdDataHistorySequence++,
                     FromStateVersion = e.StateVersion,
-                    NonFungibleIdDataId = nonFungibleIdData.Id,
+                    NonFungibleIdDefinitionId = nonFungibleIdData.Id,
                     Data = e.MutableData,
                     IsDeleted = e.IsDeleted,
                     IsLocked = e.IsLocked,
                 });
-
-                if (!nonFungibleIdStore.NonFungibleIdDataIds.Contains(nonFungibleIdData.Id))
-                {
-                    nonFungibleIdStore.NonFungibleIdDataIds.Add(nonFungibleIdData.Id);
-                }
             }
 
             void AggregateEntityResourceUsingSubstates(
@@ -1517,7 +1489,6 @@ UPDATE pending_transactions
             rowsInserted += await writeHelper.CopyEntityVaultHistory(vaultHistoryToAdd, token);
             rowsInserted += await writeHelper.CopyNonFungibleIdData(nonFungibleIdDataToAdd, token);
             rowsInserted += await writeHelper.CopyNonFungibleIdDataHistory(nonFungibleIdsMutableDataHistoryToAdd, token);
-            rowsInserted += await writeHelper.CopyNonFungibleIdStoreHistory(nonFungibleIdStoreHistoryToAdd.Values, token);
             rowsInserted += await writeHelper.CopyNonFungibleIdLocationHistory(nonFungibleIdLocationHistoryToAdd, token);
             rowsInserted += await writeHelper.CopyResourceEntitySupplyHistory(resourceEntitySupplyHistoryToAdd, token);
             rowsInserted += await writeHelper.CopyNonFungibleDataSchemaHistory(nonFungibleSchemaHistoryToAdd, token);